# ctbx-payments/docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: CTBX Payments BFF API
  version: 0.1.0-draft
  description: >-
    Contrato inicial do novo BFF. Não representa nem expõe o backend legado.
    Provedores, hosts e mecanismos criptográficos permanecem decisões internas.
servers:
  - url: https://api.example.invalid
    description: Placeholder documental; não é um endpoint utilizável
security:
  - bearerAuth: []
...
paths:
  /v1/auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/AuthResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, writeOnly: true }
      responses:
        '200': { $ref: '#/components/responses/AuthResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '204': { description: Sessão revogada ou já encerrada }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: Sessão atual
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionEnvelope' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current:
    get:
      tags: [Accounts]
      operationId: getCurrentAccount
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: Conta atual
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Account' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/balances:
    get:
      tags: [Accounts]
      operationId: getCurrentBalances
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: Saldos normalizados
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Balance' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/statement:
    get:
      tags: [Accounts]
      operationId: listStatement
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/Cursor'
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: direction, in: query, schema: { type: string, enum: [CREDIT, DEBIT] } }
      responses:
        '200': { $ref: '#/components/responses/TransactionListResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/statement/future:
    get:
      tags: [Accounts]
      operationId: listFutureTransactions
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200': { $ref: '#/components/responses/TransactionListResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/statement/blocked:
    get:
      tags: [Accounts]
      operationId: listBlockedTransactions
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200': { $ref: '#/components/responses/TransactionListResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/transactions/{id}:
    get:
      tags: [Accounts]
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Transação
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Transaction' } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/accounts/current/transactions/{id}/receipt:
    get:
      tags: [Accounts]
      operationId: getTransactionReceipt
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200': { $ref: '#/components/responses/StatementReceiptResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/keys/lookup:
    post:
      tags: [Pix]
      operationId: lookupPixKey
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses:
        '200':
          description: Consulta fictícia de chave no SANDBOX
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/PixKeyLookup' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/qr/lookup:
    post:
      tags: [Pix]
      operationId: lookupPixQr
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties: { payload: { type: string } }
      responses:
        '200':
          description: Consulta fictícia de QR reconhecido no SANDBOX
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PixQrLookup' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/keys:
    get:
      tags: [Pix]
      operationId: listPixKeys
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: Chaves da conta
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/PixKey' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Pix]
      operationId: createPixKey
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, value]
              properties:
                type: { type: string, enum: [CPF, CNPJ, EMAIL, PHONE, RANDOM] }
                value: { type: string }
      responses:
        '202': { $ref: '#/components/responses/OperationResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/keys/{id}:
    delete:
      tags: [Pix]
      operationId: deletePixKey
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/OperationResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/receive/qr:
    post:
      tags: [Pix]
      operationId: createReceivePixQr
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keyId]
              properties:
                keyId: { type: string }
                amountMinor: { type: integer, minimum: 0, description: Valor opcional em centavos inteiros }
                description: { type: string, maxLength: 140 }
      responses:
        '200':
          description: QR de recebimento
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PixReceiveQr' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/transfers/validate:
    post:
      tags: [Pix]
      operationId: validatePixTransfer
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PixValidationIntent' }
      responses:
        '200':
          description: Intenção validada, sem efetivação
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/PixValidation' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/transfers:
    post:
      tags: [Pix]
      operationId: createPixTransfer
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PixSubmitIntent' }
      responses:
        '200': { $ref: '#/components/responses/PixTransferResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/transfers/schedule:
    post:
      tags: [Pix]
      operationId: schedulePixTransfer
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PixSubmitIntent'
                - type: object
                  required: [scheduledFor]
                  properties: { scheduledFor: { type: string, format: date-time } }
      responses:
        '200': { $ref: '#/components/responses/PixTransferResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/transfers/{id}:
    get:
      tags: [Pix]
      operationId: getPixTransfer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200': { $ref: '#/components/responses/PixTransferResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /v1/pix/transfers/{id}/receipt:
    get:
      tags: [Pix]
      operationId: getPixReceipt
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Comprovante estrutural PIX SANDBOX
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/PixSandboxReceipt' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /v1/transfers/banks:
    get:
      tags: [Transfers]
      operationId: listBanks
      parameters: [{ $ref: '#/components/parameters/DeviceId' }]
      responses:
        '200':
          description: Instituições disponíveis
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Bank' } }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /v1/transfers:
    post:
      tags: [Transfers]
      operationId: createBankTransfer
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { $ref: '#/components/requestBodies/BankTransferIntent' }
      responses:
        '202': { $ref: '#/components/responses/BankTransferResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Identificador opaco devolvido no login e vinculado à sessão.
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    ErrorResponse:
      description: Erro normalizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    AuthResponse:
      description: Resultado de autenticação, inclusive estados intermediários
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthEnvelope' }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        meta: { type: object, additionalProperties: true }
        requestId: { type: string }
...
    LoginRequest:
      type: object
      required: [username, password, device]
      properties:
        username: { type: string }
        password: { type: string, writeOnly: true }
        device: { $ref: '#/components/schemas/DeviceRegistration' }
...